Updates bot to relatively stable state
I re-factored a bunch of stuff, mainly the parsing module, and this
commit brings the bot into a pretty functioning state.

This commit fixed a few problems, where the MessageInterpreter was
executing the same Action multiple times, this was solved by adding an
insertion queue mechanism, that append newly added actions to the list
of registered actions after the current batch of messages has been
processed.